requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
importlib-metadata==6.8.0
kombu==5.3.4
Markdown==3.5.1
Pillow==10.1.0
prompt-toolkit==3.0.41
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
zipp==3.17.0
